requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai==1.12.0
streamlit==1.31.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3
pdf2image==1.16.3
Pillow==10.2.0
pdfplumber==0.10.3
faiss-cpu==1.7.4
langchain==0.1.0
langchain-text-splitters==0.0.1
numpy==1.24.3
APScheduler==3.10.4
